Fixes click-outside document listener to make up for ES6 this-binding weirdness causing it to never be unbound on unmount.
The listener is instead declared in the constructor as binding this to
the handleDocument function on add/removeEventListener actually results in
a different function reference being passed to the event listener manager.

The result of that is that the click-outside listener is never actually
removed from the document because a different function reference is
passed in the request to remove than the one passed in the request to
add.